+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Unified DFS path handling when using any API that uses a transaction to open the file
   * This includes `smbclient.rename` and `smbclient.replace`
 * Fixed up `smbclient.rename` to work with directories
+* `smbclient.scandir` will continue to use the connection cache when getting stat information of a dir entry
+* `smbclient.shutil.rmtree` will continue to use the connection cache when removing child entries
 
 
 ## 1.5.0 - 2021-03-25